requirements.txt
discord.py==2.3.2
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0 